+++ resources/views/pages/s_user/karyawan/m_rembes_item/index.blade.php
@@ -12,7 +12,7 @@
     </div>
     <!-- /BREADCRUMB -->
 
-    <div class="row my-3">
+    {{-- <div class="row my-3">
         <div class="col-lg-12 margin-tb">
 
             <div class="pull-right">
@@ -22,7 +22,7 @@
                 </a>
             </div>
         </div>
-    </div>
+    </div> --}}
 
     @include('layouts.partials.alert-prompt.alert')
 
@@ -30,17 +30,11 @@
         <div class="switch align-self-center">
 
 
+
             @can('rembes-item-create')
-                @if ($rembes->status !== 'APPROVED')
+                @if ($rembes->status == 'PENDING' || ($rembes->status == 'REJECTED' && auth()->id() == $rembes->user_id))
                     <a class="btn btn-primary" href="{{ route('dashboard.rembes-item.create', $rembes->id) }}">
-                        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none"
-                            stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"
-                            class="feather feather-user-plus">
-                            <path d="M16 21v-2a4 4 0 0 0-4-4H5a4 4 0 0 0-4 4v2"></path>
-                            <circle cx="8.5" cy="7" r="4"></circle>
-                            <line x1="20" y1="8" x2="20" y2="14"></line>
-                            <line x1="23" y1="11" x2="17" y2="11"></line>
-                        </svg>
+                        <i class="far fa-plus-square"></i>
                         Add New Reimburse Item
                     </a>
                 @endif
@@ -103,36 +97,38 @@
                                     </td>
 
                                     <td>
-                                        <a class="badge badge-light-primary text-start me-2"
-                                            href="{{ route('dashboard.rembes-item.show', ['rembes' => $rembes->id, 'id' => $rembes_item->id]) }}">
-                                            <i class="far fa-edit"></i>
-                                            Detail
-                                        </a>
-                                        @can('rembes-item-edit')
-                                            @if ($rembes->status !== 'APPROVED')
-                                                <a class="badge badge-light-primary text-start me-2"
-                                                    href="{{ route('dashboard.rembes-item.edit', ['rembes' => $rembes->id, 'id' => $rembes_item->id]) }}">
-                                                    <i class="far fa-edit"></i>
-                                                    Edit
-                                                </a>
-                                            @endif
-                                        @endcan
+                                        @if ($rembes->status == 'PENDING' || ($rembes->status == 'REJECTED' && auth()->id() == $rembes->user_id))
+                                            <a class="badge badge-light-primary text-start me-2"
+                                                href="{{ route('dashboard.rembes-item.show', ['rembes' => $rembes->id, 'id' => $rembes_item->id]) }}">
+                                                <i class="far fa-edit"></i>
+                                                Detail
+                                            </a>
+                                            @can('rembes-item-edit')
+                                                @if ($rembes->status !== 'APPROVED')
+                                                    <a class="badge badge-light-primary text-start me-2"
+                                                        href="{{ route('dashboard.rembes-item.edit', ['rembes' => $rembes->id, 'id' => $rembes_item->id]) }}">
+                                                        <i class="far fa-edit"></i>
+                                                        Edit
+                                                    </a>
+                                                @endif
+                                            @endcan
 
-                                        @can('rembes-item-delete')
-                                            @if ($rembes->status !== 'APPROVED')
-                                                {!! Form::open([
-                                                    'method' => 'DELETE',
-                                                    'route' => ['dashboard.rembes-item.destroy', ['rembes' => $rembes->id, 'id' => $rembes_item->id]],
-                                                    'style' => 'display:inline',
-                                                ]) !!}
-                                                {!! Form::button('<i class="far fa-trash-alt"></i> Delete', [
-                                                    'type' => 'submit',
-                                                    'class' => 'badge badge-light-danger text-start mx-3',
-                                                    'onclick' => 'return confirm("Are you sure you want to delete this rembes item?");',
-                                                ]) !!}
-                                                {!! Form::close() !!}
-                                            @endif
-                                        @endcan
+                                            @can('rembes-item-delete')
+                                                @if ($rembes->status !== 'APPROVED')
+                                                    {!! Form::open([
+                                                        'method' => 'DELETE',
+                                                        'route' => ['dashboard.rembes-item.destroy', ['rembes' => $rembes->id, 'id' => $rembes_item->id]],
+                                                        'style' => 'display:inline',
+                                                    ]) !!}
+                                                    {!! Form::button('<i class="far fa-trash-alt"></i> Delete', [
+                                                        'type' => 'submit',
+                                                        'class' => 'badge badge-light-danger text-start mx-3',
+                                                        'onclick' => 'return confirm("Are you sure you want to delete this rembes item?");',
+                                                    ]) !!}
+                                                    {!! Form::close() !!}
+                                                @endif
+                                            @endcan
+                                        @endif
                                     </td>
                                 </tr>
                             @endforeach
